requirements.txt
autogen-agentchat
autogen-core
autogen-ext[mcp]
fastapi
uvicorn
requests
python-dotenv
tavily-python
openai
